Update the setup file with optional packages for dashboard.
