+++ RAIDashboard/certificate_page.py
@@ -17,7 +17,7 @@
 import logging
 import dash_bootstrap_components as dbc
 from dash import Input, Output, dcc, html
-from server import app, dbUtils
+from .server import app, dbUtils
 
 logger = logging.getLogger(__name__)
