+++ RAIDashboard/display_types/display_registry.py
