+++ RAIDashboard/display_types/image_element.py
