+++ RAIDashboard/setting_page.py
@@ -18,7 +18,7 @@
 import logging
 import dash_bootstrap_components as dbc
 from dash import Input, Output, html
-from server import app, dbUtils
+from .server import app, dbUtils
 import urllib
 logger = logging.getLogger(__name__)
