+++ RAIDashboard/single_metric_info_page.py
@@ -18,7 +18,7 @@
 import logging
 import dash_bootstrap_components as dbc
 from dash import html
-from server import dbUtils
+from .server import dbUtils
 logger = logging.getLogger(__name__)
